Add -config option to allow override of default config location
Allows GDA_CONFIG and GDA_INSTANCE_CONFIG_rel to be derived from the 
same absolute path specified on the command line. This also maintains 
the WORKSPACE_ROOT_LOCATION which is central to the Eclipse and exported 
versions of the server product knowing where everything is.

Also uses more robust methods of determining some paths than before.

Change-Id: I8b3fc4ef68b8a17383f3468e746881ae2ec49e68