Ensure installed build contains scanning command scripts
At the moment the scripts referenced by the various configs are part of 
the exported product since the configs require them to be relative to 
gda.install.git.loc. This probably ought to change in future so they are 
relative to the config that references them.

For the moment this ensures that the org.eclipse.scanning.command are 
available in the traditional location.

Change-Id: Ic5281b95731c50875a3571563a7c251af563052c