DAQ-241: Record server startup type for later use.
When the configuration detects whether the server has been started from 
eclipse or from an exported product, record this in a System property so 
that the rest of the application can find out later. 

Also explicitly add workspace location to Name Server startup to 
simplify trailing process detection on shutdown.

Change-Id: I91ee1431332905ea019ff05a1aaf5b13d7949297