DAQ-336 materialize extra bundles that will be required by the server
All this commit does is mean that the bundles (and dependencies) are
materialized with the server, but they are not (currently) part of the
exported product. This is because they are only used by beamlines not
yet on GDA 9, and we haven't yet worked out which feature they should
go in.

We still want to materialize them so that they are visible to developers
in the workspace, and so that test jobs can verify they compile.

Change-Id: I076a1cafb2f2d31c8267b9b982cc62cfb798bda9