Remove dls-root from server export and update config defaults
As a result of recent changes the dls-root folder that is currently
generated when exporting the GDA-9 server product is no longer required.
This change removes its creation and all the associated links from the
buckminster ant jobs and updates ConfigDefaults so that its APPS_ROOT is
always the parent folder of the workspace.

Additionally it removes the short lived create development product ant
task as this is better handled by setting and interpreting a property
in the two buckminster properties files.

Change-Id: Idde24a0038fbd7e2d99875b0d4450aec78132c1b