Move Xerces into 3rd party xml feature and rationalise feature structure
For consistency the Xerces plugin should be in the 3rd Party XML feature
which is then included by the GDA Common feature. This allows it also to
be removed from the analysis feature which (indirectly) includes the
common feature.

Also xml.resolver seems to have been in the wrong feature so this is
moved too.

Change-Id: Ib53f213162f2d6307cd61ed689ed154a8879ccd3