Add -clean to the GDA server launch arguments as default
-clean Deletes the OSGi bundle cache

If you make changes to plugins which have been cached by OSGi then if
the -clean argument is not used the changes will not be picked up. As
the performance improvement offered by this caching is minimal for the
server (as we don't restart it frequently) having -clean always applied
seems like a good idea.

See:
https://help.eclipse.org/neon/index.jsp?topic=%2Forg.eclipse.platform.doc.user%2Ftasks%2Frunning_eclipse.htm

Change-Id: I3ed4f9d97fa39e12edb4859d42e49a71defc073f