[I13-544] Use Scannable interface to address sample centring motors
DataCollectionADControllerImpl, which is used only on I13 (both
branches) is a large data structure containing (among other things)
sampleCentringXMotor & sampleCentringYMotor, which are addressed
via the interface ScannableMotionUnits and are initialised with
ScannableMotor beans on Coherence and ScannableRotatingGonio on Imaging.

This changes DataCollectionADControllerImpl to address the motors
via the Scannable interface, which both ScannableMotor and
ScannableRotatingGonio implement. This makes no practical difference,
as the only function called on them is asynchronousMoveTo(),
which is defined in Scannable.

The advantage of using Scannable is that this is the defined
@ServiceInterface for the concrete classes.

Change-Id: Ie5d1f29906baf59ec1c60e53b6b2ef7546ffbf2f