Don't try and send message to client when server startup fails
If the servers fail to start for any reason, the final error would be that it
couldn't find the command server. This added a layer of misdirection to
debugging what actually failed and as the command server was only used to
broadcast the shutdown message, it isn't needed for startup failures.

Change-Id: I13b4587a70d4812d90fdf70e234e577573e07e96