DAQ-1980 Remove LogServer from GDA
This was a separate child process spawned by the main GDA Server
application. It was useful in the past when there were multiple
servers (Corba related or otherwise).

Now the server's logging config has been modified to take on the roles
that the log server performed which was to receive messages from any
GDA clients and send messages to file and any log panels.

Change-Id: Id51d300024fd2afe4c27401f477aed6bb738db3b