DAQ-2923 Deploy built product and create symlink
Instead of switching the Maven build directory to client-product-root
unzip the generated product . The build directory contains additional
things that don't need to be deployed as the 'product'.

This is achieved by attaching the zip file (generated by the
archive-products goal) as an artifact and then unzipping to the
desired location.

The Exec Maven Plugin is used to recreate the symlink. Currently this
is disabled by default. The profile recreate-symlink can be used
to enable it.

Change-Id: I8f9119116bc3017c2fabc47c14f5d31f80bdb6d9